+++ common/config/config.hpp
@@ -13,8 +13,8 @@ auto get_repeat() -> int;
 void set_repeat(int value);
 
 // tune volume
-auto get_volume() -> float;
-void set_volume(float value);
+auto get_volume() -> double;
+void set_volume(double value);
 
 // per title tune enable
 auto has_title_enabled(u64 tid) -> bool;
@@ -27,12 +27,12 @@ void set_title_enabled_default(bool value);
 
 // per title volume
 auto has_title_volume(u64 tid) -> bool;
-auto get_title_volume(u64 tid) -> float;
-void set_title_volume(u64 tid, float value);
+auto get_title_volume(u64 tid) -> double;
+void set_title_volume(u64 tid, double value);
 
 // default volume for every title
-auto get_default_title_volume() -> float;
-void set_default_title_volume(float value);
+auto get_default_title_volume() -> double;
+void set_default_title_volume(double value);
 
 // returns true if title causes a fatal on launch
 auto get_title_blacklist(u64 tid) -> bool;
