+++ sys-tune/source/impl/music_player.hpp
@@ -26,12 +26,12 @@ namespace tune::impl {
     void Next();
     void Prev();
 
-    float GetVolume();
-    void SetVolume(float volume);
-    float GetTitleVolume();
-    void SetTitleVolume(float volume);
-    float GetDefaultTitleVolume();
-    void SetDefaultTitleVolume(float volume);
+    double GetVolume();
+    void SetVolume(double volume);
+    double GetTitleVolume();
+    void SetTitleVolume(double volume);
+    double GetDefaultTitleVolume();
+    void SetDefaultTitleVolume(double volume);
 
     void TitlePlay();
     void TitlePause();
