+++ sys-tune/source/tune_service.cpp
@@ -52,22 +52,22 @@ namespace tune {
                     return 0;
 
                 case TuneIpcCmd_GetVolume:
-                    GET_SINGLE(float, impl::GetVolume);
+                    GET_SINGLE(double, impl::GetVolume);
 
                 case TuneIpcCmd_SetVolume:
-                    SET_SINGLE(float, impl::SetVolume);
+                    SET_SINGLE(double, impl::SetVolume);
 
                 case TuneIpcCmd_GetTitleVolume:
-                    GET_SINGLE(float, impl::GetTitleVolume);
+                    GET_SINGLE(double, impl::GetTitleVolume);
 
                 case TuneIpcCmd_SetTitleVolume:
-                    SET_SINGLE(float, impl::SetTitleVolume);
+                    SET_SINGLE(double, impl::SetTitleVolume);
 
                 case TuneIpcCmd_GetDefaultTitleVolume:
-                    GET_SINGLE(float, impl::GetDefaultTitleVolume);
+                    GET_SINGLE(double, impl::GetDefaultTitleVolume);
 
                 case TuneIpcCmd_SetDefaultTitleVolume:
-                    SET_SINGLE(float, impl::SetDefaultTitleVolume);
+                    SET_SINGLE(double, impl::SetDefaultTitleVolume);
 
                 case TuneIpcCmd_GetRepeatMode:
                     GET_SINGLE(RepeatMode, impl::GetRepeatMode);
